+++ package.json
@@ -9,12 +9,14 @@
     "start": "next start"
   },
   "dependencies": {
+    "@heroicons/react": "^1.0.6",
     "@supabase/supabase-js": "^2.21.0",
     "@tanstack/react-query": "^4.28.0",
     "@trpc/client": "^10.18.0",
     "@trpc/next": "^10.18.0",
     "@trpc/react-query": "^10.18.0",
     "@trpc/server": "^10.18.0",
+    "heroicons": "^2.0.18",
     "next": "^13.2.4",
     "next-auth": "^4.21.0",
     "react": "18.2.0",
